refactor(shared): extract probe classifiers and fix sonarcloud findings
PR #36 SonarCloud surfaced 3 minor issues and ~26 lines of duplicated
code between `plugin/info.ts` and `plugin/list.ts` (the `narrowResolverNotes`
function bodies). Addressed in one pass.

Probe classifiers extracted to `shared/probe-classifiers.ts`:
- `narrowProbeError` (closed-set Reason for FS/JSON throws)
- `narrowResolverNotes` (closed-set Reason for resolver notes)

Both classifiers had three copies (`list.ts`, `plugin/info.ts`,
`marketplace/info.ts`) carrying byte-identical function bodies. The
in-source comments explicitly justified the duplication with "shared/
is the sanctioned cross-orchestrator import surface" -- which is the
exact reason to put them in shared/. Done.

Local wrappers preserved:
- `list.ts::narrowProbeError` -- thin wrapper, keeps the per-row
  semantic name at the call sites and the `__test_narrowProbeError`
  re-export stable.
- `list.ts::narrowListFailReason` -- thin wrapper, keeps the
  orchestrator-level semantic name distinct from the per-row helper.
  Same underlying ladder, two semantic names at the call sites.

Minor issue fixes (typescript rules):
- S7747 `plugin/info.ts:278` -- removed unnecessary `[...Object.keys()]`
  spread; `Object.keys()` already returns a fresh array.
- S7735 `notify.ts:1803` -- inverted ternary so the positive branch is
  the if-clause (refSuffix github source).
- S7735 `notify.ts:1996` -- inverted ternary (plugin.scope bracket).

Net -149 LOC. All 1464 tests pass; npm run check clean.
